# controllers/AuthController.go
package controllers

import (
	"Final-Project-JCC-Golang-2022/models"
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

type LoginInput struct {
    Username string `json:"username" binding:"required"`
    Password string `json:"password" binding:"required"`
}

type RegisterInput struct {
    Username string `json:"username" binding:"required"`
    Password string `json:"password" binding:"required"`
    Email    string `json:"email" binding:"required"`
}

// LoginAdmin godoc
// @Summary Login as as admin.
// @Description Logging in to get jwt token to access admin or user api by roles. Admin can add or update product, product category, and payment method
// @Tags Auth
// @Param Body body LoginInput true "the body to login a admin"
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Router /login-admin [post]
func Login(c *gin.Context) {
    db := c.MustGet("db").(*gorm.DB)
    var input LoginInput

    if err := c.ShouldBindJSON(&input); err != nil {
        c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
        return
    }

    u := models.Admin{}

    u.Username = input.Username
    u.Password = input.Password

    token, err := models.LoginCheck(u.Username, u.Password, db)

    if err != nil {
        fmt.Println(err)
        c.JSON(http.StatusBadRequest, gin.H{"error": "username or password is incorrect."})
        return
    }

    admin := map[string]string{
        "username": u.Username,
        "email":    u.Email,
    }

    c.JSON(http.StatusOK, gin.H{"message": "login success", "user": admin, "token": token})

}

// Register godoc
// @Summary Register a admin.
// @Description registering a admin from public access.
// @Tags Auth
// @Param Body body RegisterInput true "the body to register a admin"
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Router /register-admin [post]
func Register(c *gin.Context) {
    db := c.MustGet("db").(*gorm.DB)
    var input RegisterInput

    if err := c.ShouldBindJSON(&input); err != nil {
        c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
        return
    }

    u := models.Admin{}

    u.Username = input.Username
    u.Email = input.Email
    u.Password = input.Password

    _, err := u.SaveAdmin(db)

    if err != nil {
        c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
        return
    }

    admin := map[string]string{
        "username": input.Username,
        "email":    input.Email,
    }

    c.JSON(http.StatusOK, gin.H{"message": "registration success", "admin": admin})

}